Restructure review page: move approvals and actions into a right sidebar
The status chip, approval progress, and run/menu actions now live in a
page-level sidebar instead of inside the request boxes, so the SQL box
gets the full content column width. Description, statement, results and
activity timeline share one consistent column. Code blocks render at
13px and get themed thin scrollbars via the tailwind-scrollbar plugin.
