+++ frontend/src/routes/Review/index.tsx
@@ -5,6 +5,8 @@ import Spinner from "../../components/Spinner";
 import useRequest from "../../hooks/request";
 import KubernetesRequestDisplay from "./KubernetesRequestDisplay";
 import DatasourceRequestDisplay from "./DatasourceRequestDisplay";
+import DatasourceRequestSidebar from "./DatasourceRequestSidebar";
+import KubernetesRequestSidebar from "./KubernetesRequestSidebar";
 import ActivityTimeline from "./ActivityTimeline";
 import NotAuthorized from "../../components/NotAuthorized";
 
@@ -51,43 +53,47 @@ function RequestReview() {
                 { label: request.title },
               ]}
             />
-            <h1 className="my-2 flex w-full items-start text-3xl">
-              <div className="mr-auto">{request?.title}</div>
-              <div
-                className={` ${mapStatusToLabelColor(
-                  mapStatus(request.reviewStatus, request.executionStatus),
-                )} mt-2 rounded-md px-2 py-1 text-base font-medium ring-1 ring-inset `}
-              >
-                {mapStatus(request.reviewStatus, request.executionStatus)}
-              </div>
-            </h1>
-            <div className="">
-              <div className="">
-                {request &&
-                  (request?._type === "DATASOURCE" ? (
-                    <DatasourceRequestDisplay
-                      request={request}
-                      run={run}
-                      cancelQuery={cancelQuery}
-                      start={start}
-                      updateRequest={updateRequest}
-                      results={results}
-                      dataLoading={dataLoading}
-                      executionError={executionError}
-                      proxyResponse={proxyResponse}
-                    ></DatasourceRequestDisplay>
-                  ) : (
-                    <KubernetesRequestDisplay
-                      request={request}
-                      run={run}
-                      start={start}
-                      updateRequest={updateRequest}
-                      results={kubernetesResults}
-                      dataLoading={dataLoading}
-                      executionError={executionError}
-                      proxyResponse={proxyResponse}
-                    ></KubernetesRequestDisplay>
-                  ))}
+            <h1 className="my-2 text-3xl">{request?.title}</h1>
+            <div className="flex flex-col gap-6 md:flex-row md:items-start">
+              <aside className="flex w-full flex-col gap-4 border-slate-200 dark:border-slate-700 md:order-last md:w-60 md:shrink-0 md:border-l md:pl-4">
+                <div
+                  className={`${mapStatusToLabelColor(
+                    mapStatus(request.reviewStatus, request.executionStatus),
+                  )} w-fit rounded-md px-2 py-1 text-sm font-medium ring-1 ring-inset`}
+                >
+                  {mapStatus(request.reviewStatus, request.executionStatus)}
+                </div>
+                {request._type === "DATASOURCE" ? (
+                  <DatasourceRequestSidebar
+                    request={request}
+                    runQuery={run}
+                    cancelQuery={cancelQuery}
+                    startServer={start}
+                  />
+                ) : (
+                  <KubernetesRequestSidebar request={request} runQuery={run} />
+                )}
+              </aside>
+              <div className="min-w-0 flex-1">
+                {request._type === "DATASOURCE" ? (
+                  <DatasourceRequestDisplay
+                    request={request}
+                    updateRequest={updateRequest}
+                    results={results}
+                    dataLoading={dataLoading}
+                    executionError={executionError}
+                    proxyResponse={proxyResponse}
+                  ></DatasourceRequestDisplay>
+                ) : (
+                  <KubernetesRequestDisplay
+                    request={request}
+                    updateRequest={updateRequest}
+                    results={kubernetesResults}
+                    dataLoading={dataLoading}
+                    executionError={executionError}
+                    proxyResponse={proxyResponse}
+                  ></KubernetesRequestDisplay>
+                )}
                 <div className="mt-3 w-full border-b border-slate-300 dark:border-slate-700"></div>
                 <ActivityTimeline
                   request={request}
